errors: move ConnectionSetupRequestError to scylla crate
