+++ scylla/src/errors.rs
@@ -2,12 +2,14 @@ use std::{io::ErrorKind, sync::Arc};
 
 use scylla_cql::{
     errors::{
-        BadKeyspaceName, BadQuery, BrokenConnectionError, ConnectionSetupRequestError,
-        CqlResponseKind, DbError, RequestError, TranslationError,
+        BadKeyspaceName, BadQuery, BrokenConnectionError, CqlRequestKind, CqlResponseKind, DbError,
+        RequestError, TranslationError,
     },
     frame::{
         frame_errors::{
-            CqlErrorParseError, CqlResponseParseError, CqlResultParseError, FrameError, ParseError,
+            CqlAuthChallengeParseError, CqlAuthSuccessParseError, CqlAuthenticateParseError,
+            CqlErrorParseError, CqlResponseParseError, CqlResultParseError, CqlSupportedParseError,
+            FrameError, ParseError,
         },
         value::SerializeValuesError,
     },
@@ -19,7 +21,7 @@ use scylla_cql::{
 
 use thiserror::Error;
 
-use crate::frame::response;
+use crate::{authentication::AuthError, frame::response};
 
 /// Error that occurred during query execution
 #[derive(Error, Debug, Clone)]
@@ -290,6 +292,64 @@ impl ConnectionError {
         }
 
         false
+    }
+}
+
+/// An error that occurred during connection setup request execution.
+/// It indicates that request needed to initiate a connection failed.
+#[derive(Error, Debug, Clone)]
+#[error("Failed to perform a connection setup request. Request: {request_kind}, reason: {error}")]
+pub struct ConnectionSetupRequestError {
+    request_kind: CqlRequestKind,
+    error: ConnectionSetupRequestErrorKind,
+}
+
+#[derive(Error, Debug, Clone)]
+#[non_exhaustive]
+pub enum ConnectionSetupRequestErrorKind {
+    // TODO: Make FrameError clonable.
+    #[error(transparent)]
+    FrameError(Arc<FrameError>),
+    #[error("Unable to allocate stream id")]
+    UnableToAllocStreamId,
+    #[error(transparent)]
+    BrokenConnection(#[from] BrokenConnectionError),
+    #[error("Database returned an error: {0}, Error message: {1}")]
+    DbError(DbError, String),
+    #[error("Received unexpected response from the server: {0}")]
+    UnexpectedResponse(CqlResponseKind),
+    #[error("Failed to deserialize SUPPORTED response: {0}")]
+    CqlSupportedParseError(#[from] CqlSupportedParseError),
+    #[error("Failed to deserialize AUTHENTICATE response: {0}")]
+    CqlAuthenticateParseError(#[from] CqlAuthenticateParseError),
+    #[error("Failed to deserialize AUTH_SUCCESS response: {0}")]
+    CqlAuthSuccessParseError(#[from] CqlAuthSuccessParseError),
+    #[error("Failed to deserialize AUTH_CHALLENGE response: {0}")]
+    CqlAuthChallengeParseError(#[from] CqlAuthChallengeParseError),
+    #[error("Failed to deserialize ERROR response: {0}")]
+    CqlErrorParseError(#[from] CqlErrorParseError),
+    #[error("Failed to start client's auth session: {0}")]
+    StartAuthSessionError(AuthError),
+    #[error("Failed to evaluate auth challenge on client side: {0}")]
+    AuthChallengeEvaluationError(AuthError),
+    #[error("Failed to finish auth challenge on client side: {0}")]
+    AuthFinishError(AuthError),
+    #[error("Authentication is required. You can use SessionBuilder::user(\"user\", \"pass\") to provide credentials or SessionBuilder::authenticator_provider to provide custom authenticator")]
+    MissingAuthentication,
+}
+
+impl From<FrameError> for ConnectionSetupRequestErrorKind {
+    fn from(value: FrameError) -> Self {
+        ConnectionSetupRequestErrorKind::FrameError(Arc::new(value))
+    }
+}
+
+impl ConnectionSetupRequestError {
+    pub fn new(request_kind: CqlRequestKind, error: ConnectionSetupRequestErrorKind) -> Self {
+        ConnectionSetupRequestError {
+            request_kind,
+            error,
+        }
     }
 }
