+++ scylla/src/transport/connection.rs
@@ -1,9 +1,8 @@
 use bytes::Bytes;
 use futures::{future::RemoteHandle, FutureExt};
 use scylla_cql::errors::{
-    BrokenConnectionError, BrokenConnectionErrorKind, ConnectionSetupRequestError,
-    ConnectionSetupRequestErrorKind, CqlEventHandlingError, CqlRequestKind, RequestError,
-    ResponseParseError, TranslationError,
+    BrokenConnectionError, BrokenConnectionErrorKind, CqlEventHandlingError, CqlRequestKind,
+    RequestError, ResponseParseError, TranslationError,
 };
 use scylla_cql::frame::frame_errors::CqlResponseParseError;
 use scylla_cql::frame::request::options::{self, Options};
@@ -34,7 +33,10 @@ use tokio_openssl::SslStream;
 pub(crate) use ssl_config::SslConfig;
 
 use crate::authentication::AuthenticatorProvider;
-use crate::errors::{ConnectionError, QueryError, UserRequestError};
+use crate::errors::{
+    ConnectionError, ConnectionSetupRequestError, ConnectionSetupRequestErrorKind, QueryError,
+    UserRequestError,
+};
 use scylla_cql::frame::response::authenticate::Authenticate;
 use std::collections::{BTreeSet, HashMap, HashSet};
 use std::convert::TryFrom;
